+++ src/main/java/com/bookshop/controllers/UserController.java
@@ -1,8 +1,5 @@
 package com.bookshop.controllers;
 
-import java.util.List;
-import java.util.Optional;
-
 import com.bookshop.dao.User;
 import com.bookshop.dto.UserDTO;
 import com.bookshop.exceptions.DuplicateRecordException;
@@ -11,7 +8,6 @@ import com.bookshop.exceptions.NotFoundException;
 import com.bookshop.helpers.CheckValid;
 import com.bookshop.helpers.ConvertObject;
 import com.bookshop.repositories.UserRepository;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -19,15 +15,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping(value = "/api/users")
@@ -41,7 +32,8 @@ public class UserController {
 
     @GetMapping
     @PreAuthorize("@userAuthorizer.authorizeAdmin(authentication, 'ADMIN')")
-    public ResponseEntity<?> getAllUsers(@RequestParam(name = "page", required = false) Integer pageNum) {
+    public ResponseEntity<?> getAllUsers(@RequestParam(name = "page", required = false) Integer pageNum,
+                                         @RequestParam(name = "search", required = false) String search) {
         if (pageNum != null) {
             Page<User> page = userRepository.findAll(PageRequest.of(pageNum.intValue(), 10));
             if (page.getNumberOfElements() == 0) {
@@ -49,6 +41,15 @@ public class UserController {
             }
             return ResponseEntity.ok().body(page.getContent());
         }
+
+        if (search != null) {
+            List<User> usersSearch = userRepository.findByUsernameContaining(search);
+            if (usersSearch.size() == 0) {
+                return ResponseEntity.noContent().build();
+            }
+            return ResponseEntity.ok().body(usersSearch);
+        }
+
         List<User> users = userRepository.findAll();
         if (users.size() == 0) {
             return ResponseEntity.noContent().build();
@@ -80,6 +81,30 @@ public class UserController {
             User newUser = userRepository.save(user);
 
             return ResponseEntity.status(HttpStatus.CREATED).body(newUser);
+        }
+        throw new InvalidException("Invalid user");
+    }
+
+    @PutMapping("/{userId}")
+    @PreAuthorize("@userAuthorizer.authorizeUser(authentication, #userId)")
+    public ResponseEntity<?> editUserClient(@RequestBody UserDTO userDTO, @PathVariable("userId") Long userId) {
+        Optional<User> optionalUser = userRepository.findById(userId);
+        if (!optionalUser.isPresent()) {
+            throw new NotFoundException("User not found");
+        }
+        User user = optionalUser.get();
+
+        if (CheckValid.checkUpdateProfile(userDTO)) {
+
+            user.setFirstName(userDTO.getFirstName().trim().replaceAll("\\s+", " "));
+            user.setLastName(userDTO.getLastName().trim().replaceAll("\\s+", " "));
+            user.setEmail(userDTO.getEmail());
+            user.setPhone(userDTO.getPhone());
+            user.setAddress(userDTO.getAddress().trim().replaceAll("\\s+", " "));
+
+            User updatedUser = userRepository.save(user);
+
+            return ResponseEntity.status(HttpStatus.OK).body(updatedUser);
         }
         throw new InvalidException("Invalid user");
     }
@@ -129,6 +154,11 @@ public class UserController {
         if (!optionalUser.isPresent()) {
             throw new NotFoundException("User not found");
         }
+
+        if (!optionalUser.get().getSaleOrders().isEmpty()) {
+            throw new InvalidException("Delete failed");
+        }
+
         userRepository.deleteById(userId);
 
         return ResponseEntity.status(HttpStatus.OK).body(optionalUser.get());
